requirements.txt
appnope==0.1.4
asttokens==2.4.1
boto3==1.29.1
botocore==1.32.1
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
comm==0.2.1
cryptography==42.0.5
debugpy==1.6.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
greenlet==3.0.3
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.29.3
ipython==8.9.0
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
nest_asyncio==1.6.0
numexpr==2.8.5
numpy==1.24.3
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pip==23.3.1
platformdirs==4.2.0
prompt-toolkit==3.0.42
psutil==5.9.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
s3transfer==0.7.0
setuptools==68.2.2
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.2
tabula-py==2.9.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.13
wheel==0.41.2
zipp==3.17.0
